requirements.txt
bilibili-api-python

